refactor(generate): use dynamic import for subcommand
