+++ packages/opencode/src/cli/cmd/generate/handler.ts
@@ -1,0 +1,48 @@
+import { Server } from "../../../server/server"
+import { PublicApi } from "../../../server/routes/instance/httpapi/public"
+import { OpenApi } from "effect/unstable/httpapi"
+import type { GenerateArgs } from "./command"
+
+export async function handler(args: GenerateArgs) {
+  const specs = args.httpapi ? OpenApi.fromApi(PublicApi) : await Server.openapi()
+  for (const item of Object.values(specs.paths)) {
+    for (const method of ["get", "post", "put", "delete", "patch"] as const) {
+      const operation = item[method]
+      if (!operation?.operationId) continue
+      operation["x-codeSamples"] = [
+        {
+          lang: "js",
+          source: [
+            `import { createOpencodeClient } from "@opencode-ai/sdk`,
+            ``,
+            `const client = createOpencodeClient()`,
+            `await client.${operation.operationId}({`,
+            `  ...`,
+            `})`,
+          ].join("\n"),
+        },
+      ]
+    }
+  }
+  const raw = JSON.stringify(specs, null, 2)
+
+  // Format through prettier so output is byte-identical to committed file
+  // regardless of whether ./script/format.ts runs afterward.
+  const prettier = await import("prettier")
+  const babel = await import("prettier/plugins/babel")
+  const estree = await import("prettier/plugins/estree")
+  const format = prettier.format ?? prettier.default?.format
+  const json = await format(raw, {
+    parser: "json",
+    plugins: [babel.default ?? babel, estree.default ?? estree],
+    printWidth: 120,
+  })
+
+  // Wait for stdout to finish writing before process.exit() is called
+  await new Promise<void>((resolve, reject) => {
+    process.stdout.write(json, (err) => {
+      if (err) reject(err)
+      else resolve()
+    })
+  })
+}
